requirements.txt
asgiref==3.8.1
Django==5.1.7
django-redis==5.4.0
djangorestframework==3.16.0
psycopg2-binary==2.9.10
redis==5.2.1
sqlparse==0.5.3
